New comms for message transmit with server code
This upload contains the full project for sending to the server one message broken, which the server will break down into its parts. Our messages consist of 8 fields:

1) START Byte: defined as 0x8f -- no ASCII characters share this value so is guaranteed to be unique
2) MTYPE Byte: defined as 0x01 for post (update server), 0x02 for reply (request info), 0x04 for acknowledge (recognize message transfer), and 0x08 for nack (recognize attempt but not data). We have 4 more values available here if we need it.
3) SEQUE Byte: defined as the # of bytes in payload -- nifty for parsing and so we know when we're done.
4) SOTex Byte: defined as the ASCII character for the start of text (0x02)
5) DATAX Byte: defined as the bytes allocated for the payload
6) EOTex Byte: defined as the ASCII character for the end of text (0x03)
7) CHECK Byte: defined as the XOR of all values in the payload -- makes sure the data is right, if only loosely
8) END__ Byte: define as 0xCF -- no ASCII characters share this value so is guaranteed to be unique

When you run this you must have WiFly properly configured and the Soft AP module running, AS WELL AS the server.
Other notes, while this code does implement message TX with interrupts, it does not implement the queue. Also, in App_Tasks(), I have a rudimentary state machine which has a count and a Deinitialize call. This is intentional so that the program does not endlessly spit out the same message. Should we decide to implement a more robust state machine that part can safely be taken out. Lastly, the server is by no means robust. It was designed purely with the intention of easing the debugging process and checking for the presence of all components of the message outlined above. It cannot connect to more than one client and should not be used for demo purposes. Feel free to pull from this all you like, and please familiarize yourselves with our newly christened message format.

Things we still need to put in, but will be relevant later, are things such as what key-value pairs we want to use to signify events in the JSON payload. Also, I think it will be a good idea to use the JSMN framework, rather than trying to port our own C JSON parser. And finally, we'll need to implement what I have here with the queue system in FreeRTOS, on both the TX and RX side. From here we'll be able to do actual automated testing and thus, MS 3.